requirements.txt
tensorflow_gpu==1.11.0
internetarchive==1.8.1
numpy==1.15.1
scipy==1.1.0
Pillow==5.3.0
tables==3.4.4
PyYAML==3.13
